Added more debug visualizations. Especially for forces. Added ability to switch on/off different sources of tourqe, as well as various debug visualizations from in-engine.
Discovered that there is a large difference in the VF forces being applied in unity and C++. The ones in C++ seems to grow infinitely large (downwards), the ones in Unity seem to more correctly match the error in orientation needed to be adressed. Can this be due to the currently larger PD gain in the LF in c++? Either way, the direction (down) still seem off in that case.

 I've also made the feet huger and uniform, in order to make the model overall stabler, while I try to find the source of the absolute non-stiffness in the legs.

Creating a heavier LF results in stronger corrective movement in the legs. Probably as it then forms a firmer base for applying the GC VFs.